more work on part b

+++ HW1/tokens.hpp
@@ -41,6 +41,6 @@ enum tokentype {
 
 extern int yylineno;
 extern char* yytext;
-extern size_t yyleng;
+extern int yyleng;
 extern int yylex();
 #endif /* TOKENS_HPP_ */
